Change ushort to uint. Change back matrix interface.  (#40)
* Abstract adjMatrix into Matrix. Implement SparseMatrix.

* Change matrix interface.

* Update SparseMatrix.hpp

* Fix indentation.

* Add SparseMatrix multiply function copyright.

* Update matrix interface when WEIGHTED is turned on

* remove redundant

* Change sparse matrix implementation

* Fix sparse matrix index usage.

* Change all ushort to uint.

* Rename VALUE_UNIT to WEIGHTED_VALUE.

* Add template support for class Matrix.

* Optimize initConnectedComponents function.

* Abstract CORES' data structure into Matrix.

* Disable printing measure 'ses' when SPARSE is enabled.

* Fix SparseMatrix multiply index.

* Change matrix's get set method back to brackets.

* Rename WEIGHTED_VALUE to MATRIX_UNIT.

+++ src/arguments/graphLoader.cpp
@@ -284,8 +284,8 @@ void initGraphs(Graph& G1, Graph& G2, ArgumentParser& args) {
 
         vector<string> validLocksG1;
         vector<string> validLocksG2;
-        unordered_map<string,ushort> mapG1 = G1.getNodeNameToIndexMap();
-        unordered_map<string,ushort> mapG2 = G2.getNodeNameToIndexMap();
+        unordered_map<string,uint> mapG1 = G1.getNodeNameToIndexMap();
+        unordered_map<string,uint> mapG2 = G2.getNodeNameToIndexMap();
 
         for(uint i = 0; i < column1.size(); i++){
             bool validLock = true;
