+++ src/measures/localMeasures/EdgeDensity.hpp
@@ -4,18 +4,18 @@
 
 class EdgeDensity: public LocalMeasure {
 public:
-    EdgeDensity(Graph* G1, Graph* G2, ushort maxDist);
+    EdgeDensity(Graph* G1, Graph* G2, uint maxDist);
     virtual ~EdgeDensity();
 private:
     void initSimMatrix();
     float compare(double n1, double n2);
-    double calcEdgeDensity(vector<vector<ushort> > adjList, ushort originNode, ushort numNodes, ushort maxDist);
-    vector<double> generateVector(Graph* g, ushort maxDist);
-    vector<ushort> numEdgesAround(ushort node, ushort maxDist) const;
+    double calcEdgeDensity(vector<vector<uint> > adjList, uint originNode, uint numNodes, uint maxDist);
+    vector<double> generateVector(Graph* g, uint maxDist);
+    vector<uint> numEdgesAround(uint node, uint maxDist) const;
     vector<double> edged1;
     vector<double> edged2;
-    vector<vector<ushort> > edgeList; //edges in no particular order
-    ushort maxDist;
+    vector<vector<uint> > edgeList; //edges in no particular order
+    uint maxDist;
 };
 
 #endif
