+++ src/methods/wrappers/MIGRAALWrapper.cpp
@@ -33,9 +33,9 @@ string MIGRAALWrapper::generateAlignment() {
 
 Alignment MIGRAALWrapper::loadAlignment(Graph* G1, Graph* G2, string fileName) {
     vector<string> words = fileToStrings(fileName);
-    vector<ushort> mapping(G1->getNumNodes(), G2->getNumNodes());
-    unordered_map<string, ushort> g1nodeMap = G1->getNodeNameToIndexMap();
-    unordered_map<string, ushort> g2nodeMap = G2->getNodeNameToIndexMap();
+    vector<uint> mapping(G1->getNumNodes(), G2->getNumNodes());
+    unordered_map<string, uint> g1nodeMap = G1->getNodeNameToIndexMap();
+    unordered_map<string, uint> g2nodeMap = G2->getNodeNameToIndexMap();
 
     for (uint i = 0; i < words.size(); i+=2) {
         string node1 = words[i];
